+++ RELEASENOTES.md
@@ -138,6 +138,10 @@
     *   Remove deprecated
         `DefaultLoadControl.Builder.createDefaultLoadControl()`, use `build()`
         instead.
+    *   Remove deprecated `MediaItem.PlaybackProperties`, use
+        `MediaItem.LocalConfiguration` instead. Deprecated field
+        `MediaItem.playbackProperties` is now of type
+        `MediaItem.LocalConfiguration`.
 
 ## 1.1